Updated SoundPool API for M4A 4.0

+++ MonoGame.Framework/Android/Audio/Sound.cs
@@ -9,7 +9,7 @@ namespace Microsoft.Xna.Framework.Audio
     internal class Sound : IDisposable
     {
         private const int MAX_SIMULTANEOUS_SOUNDS = 10;
-        private static SoundPool s_soundPool = new SoundPool(MAX_SIMULTANEOUS_SOUNDS, (int)Stream.Music, 0);
+        private static SoundPool s_soundPool = new SoundPool(MAX_SIMULTANEOUS_SOUNDS, Stream.Music, 0);
         private int _soundId;
         private int _streamId;
